// src/app/core/models/i-personadata.ts
import { IPersona } from "./i-persona";

export interface IPersonadata extends IPersona{
	// apellido: string;
    // nombre: string;
	// clave: string;
	// correo: string;
	// codGenero: string;
    // fechaNacimiento: string;
	// codNacionalidad: string;
	equipos: number[];
	actividades: number[];
}